synapticsmst; Partially rewrite the plugin
This removes the 'two-layer' FuDevice and FuSyanpticsmstDevice model, where
a complicated plugin cache was used to add and remove devices to the daemon.

This patch instead just adds all MST devices even if there is no Synaptics MST
controller active. The UPDATABLE flag is added when the device is powered on,
and then shows up in the system UI. This means we can remove an abstraction
layer and make everything simpler.

By making FuSynapticsmstDevice derive from FuDevice rather than GObject we can
also use a lot of the helper functionality like the other plugins, for instance
->prepare_firmware().

THIS IS WIP FOR INITIAL COMMENTS AND HAS NOT BEEN TESTED. DO NOT MERGE...